+++ jest.config.js
@@ -24,6 +24,7 @@ module.exports = {
     '^lodash-es$': 'lodash',
     '^@ovh-ux/shell$': '<rootDir>/packages/components/ovh-shell/src',
     '^@ovh-ux/ovh-reket$': '<rootDir>/packages/components/ovh-reket/src',
+    '^@ovh-ux/url-builder$': '<rootDir>/packages/manager/core/url-builder/src',
     '^@ovh-ux/manager-config$': '<rootDir>/packages/manager/modules/config/src',
   },
   watchPlugins: [