+++ packages/manager/tools/registry/src/index.js
@@ -1,9 +1,0 @@
-const dev = require('./dev');
-const generateManifests = require('./generateManifests');
-const serveStatic = require('./serveStatic');
-
-module.exports = {
-  dev,
-  generateManifests,
-  serveStatic,
-};